Features: 1.Fix a bit the display of Execution history 2. Remove BaseViewModel 3. Add common execution worker 4. Add Exection pages for exected flow steps. 5. change acuracy from double to decimal 6. Save result image to json 7. Fix orientation of result rectangle 8. Select treeview item on execution selection

+++ Model/Models/FlowStep.cs
@@ -53,7 +53,7 @@ namespace Model.Models
         public string _templateImagePath = "";
 
         [ObservableProperty]
-        public double _accuracy = 0.00d;
+        public decimal _accuracy = 0.00m;
 
         public Point ResultLocation { get; set; }
